Backend/requirements/base.txt
asgiref==3.8.1
Django==5.0.6
django-crontab==0.7.1
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
numpy==1.26.4
PyJWT==2.8.0
scikit-learn==1.3.0
scipy==1.13.1
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1